+++ src/main/scala/org/ergoplatform/network/message/BasicMessagesRepo.scala
@@ -1,6 +1,5 @@
 package org.ergoplatform.network.message
 
-import org.ergoplatform.modifiers.{ErgoNodeViewModifier, NetworkObjectTypeId}
 import org.ergoplatform.network.{PeerSpec, PeerSpecSerializer}
 import org.ergoplatform.nodeView.state.SnapshotsInfo
 import org.ergoplatform.nodeView.state.UtxoState.{ManifestId, SubtreeId}
@@ -9,78 +8,11 @@ import org.ergoplatform.network.message.MessageConstants.MessageCode
 import scorex.crypto.hash.Digest32
 import scorex.util.Extensions._
 import scorex.util.serialization.{Reader, Writer}
-import scorex.util.{ModifierId, ScorexLogging, bytesToId, idToBytes}
+import scorex.util.ModifierId
 import org.ergoplatform.sdk.wallet.Constants.ModifierIdLength
-
-import scala.collection.immutable
-
-/**
-  * Wrapper for block sections of the same type. Used to send multiple block sections at once ove the wire.
-  */
-case class ModifiersData(typeId: NetworkObjectTypeId.Value, modifiers: Map[ModifierId, Array[Byte]])
 
 case class NipopowProofData(m: Int, k: Int, headerId: Option[ModifierId]) {
   def headerIdBytesOpt: Option[Array[Byte]] = headerId.map(Algos.decode).flatMap(_.toOption)
-}
-
-/**
-  * The `Modifier` message is a reply to a `RequestModifier` message which requested these modifiers.
-  */
-object ModifiersSpec extends MessageSpecV1[ModifiersData] with ScorexLogging {
-
-  val maxMessageSize: Int = 2048576
-
-  private val maxMsgSizeWithReserve = maxMessageSize * 4 // due to big ADProofs
-
-  override val messageCode: MessageCode = 33: Byte
-  override val messageName: String = "Modifier"
-
-  private val HeaderLength = 5 // msg type Id + modifiersCount
-
-  override def serialize(data: ModifiersData, w: Writer): Unit = {
-
-    val typeId = data.typeId
-    val modifiers = data.modifiers
-    require(modifiers.nonEmpty, "empty modifiers list")
-
-    val (msgCount, msgSize) = modifiers.foldLeft((0, HeaderLength)) { case ((c, s), (_, modifier)) =>
-      val size = s + ErgoNodeViewModifier.ModifierIdSize + 4 + modifier.length
-      val count = if (size <= maxMsgSizeWithReserve) c + 1 else c
-      count -> size
-    }
-
-    w.put(typeId)
-    w.putUInt(msgCount)
-
-    modifiers.take(msgCount).foreach { case (id, modifier) =>
-      w.putBytes(idToBytes(id))
-      w.putUInt(modifier.length)
-      w.putBytes(modifier)
-    }
-
-    if (msgSize > maxMsgSizeWithReserve) {
-      log.warn(s"Message with modifiers ${modifiers.keySet} has size $msgSize exceeding limit $maxMsgSizeWithReserve.")
-    }
-  }
-
-  override def parse(r: Reader): ModifiersData = {
-    val typeId = NetworkObjectTypeId.fromByte(r.getByte()) // 1 byte
-    val count = r.getUInt().toIntExact // 8 bytes
-    require(count > 0, s"Illegal message with 0 modifiers of type $typeId")
-    val resMap = immutable.Map.newBuilder[ModifierId, Array[Byte]]
-    (0 until count).foldLeft(HeaderLength) { case (msgSize, _) =>
-      val id = bytesToId(r.getBytes(ErgoNodeViewModifier.ModifierIdSize))
-      val objBytesCnt = r.getUInt().toIntExact
-      val newMsgSize = msgSize + ErgoNodeViewModifier.ModifierIdSize + objBytesCnt
-      if (newMsgSize > maxMsgSizeWithReserve) { // buffer for safety
-        throw new Exception("Too big message with modifiers, size: " + maxMsgSizeWithReserve)
-      }
-      val obj = r.getBytes(objBytesCnt)
-      resMap += (id -> obj)
-      newMsgSize
-    }
-    ModifiersData(typeId, resMap.result())
-  }
 }
 
 /**
